Fix stock card balance rebuild after transfer dispatch, race-safe RIS number generation, and add requisition_items warehouse index
- Rebuild running stock-card balances for all touched items after stock transfer dispatch to prevent balance drift on partial dispatches
- Generate RIS numbers using a MySQL advisory lock (GET_LOCK) to prevent duplicate numbers under concurrent requests
- Add named index ri_warehouse_id on requisition_items.warehouse_id for warehouse-scoped RIS queries
- Refactor report/RIS/transfer logic and cascade service for consistency
- Update related feature tests (delivery subsidy, inventory reversal, stock transfer corrections)
- Organize documentation reports from project root into docs/
- Add RFQA-WGIMS-Updated.md QA assessment document

+++ app/Models/RequisitionItem.php
@@ -10,7 +10,7 @@ class RequisitionItem extends Model
     // are stored ONLY on requisition_dispatch_items — never on requisition_items.
     // A requisition item represents a REQUEST, not a dispatch allocation.
     protected $fillable = ['requisition_id', 'catalog_item_id', 'item_id', 'description', 'unit', 'account_code', 'warehouse_id', 'quantity_requested', 'quantity_issued', 'stock_available', 'remarks'];
-    protected $casts = ['quantity_requested' => 'integer', 'quantity_issued' => 'integer', 'stock_available' => 'boolean'];
+    protected $casts = ['quantity_requested' => 'float', 'quantity_issued' => 'float', 'stock_available' => 'boolean'];
 
     public function requisition()
     {
